+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Function get_relationship() to retrieve the relationship object.
 - Function get_model() to retrieve the model schema.
 - Function get_field() to retrieve the model field schema.
+- New "Add New Entry" option in content model dropdowns.
 
 ### Changed
 - Pressing enter in a repeating number, date or single line text field now moves focus to the next field, or adds a new field if focus is in the last field.